move native_execution_task to presto-main
