+++ presto-main/src/main/java/com/facebook/presto/native_execution/ForNativeExecutionTask.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.native_execution;
 
 import javax.inject.Qualifier;
