+++ presto-main/src/test/java/com/facebook/presto/native_execution/TestNativeExecutionTask.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.native_execution;
 
 import com.facebook.airlift.bootstrap.Bootstrap;
 import com.facebook.airlift.http.client.testing.TestingHttpClient;
